test: move the file to plugin directory to test all plugin types
